backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-magic==0.4.27
zipfile36==0.1.3
pathlib2==2.3.7
python-dotenv==1.0.0
openai==1.101.0
Flask==2.3.3